pyproject.toml
[project]
name = "deriv-api-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.2.0",
    "python-deriv-api>=0.1.6",
    "python-dotenv>=1.0.1",
]
